move runner to main package
